Implement proper speed up with 1-9 whereas 9 is as fast as possible and 1 is 0.25 speed, 3 is normal speed. 0 is as fast as possible without graphical input and only limited keyboard input (you can only turn the speed back to 1-9). Closes #127.

+++ Minemonics/src/model/universe/environments/physics/PhysicsController.hpp
@@ -134,7 +134,7 @@ private:
 	/**
 	 * The simulation speed
 	 */
-	double mSimulationSpeed;
+	int mSimulationSpeed;
 };
 
 #endif /* MODEL_UNIVERSE_ENVIRONMENTS_PHYSICS_PHYSICSCONTROLLER_H_ */
